+++ test/web/frontend/static/general.js
@@ -1,0 +1,3 @@
+window.setTimeout(() => {
+    alert('hhhh');
+}, 5);