+++ test/web/index.php
@@ -21,8 +21,17 @@ $web_service = Ark()->webService();
 //$web_service->setLogger($logger);
 $web_service->setLogger(new \sinri\ark\core\ArkLogger(__DIR__ . '/../log', 'web'));
 $router = $web_service->getRouter();
-//$router->setDebug(true);
-//$router->setLogger($logger);
+$router->setDebug(true);
+$router->setLogger($logger);
+
+$router->setErrorHandler(\sinri\ark\web\ArkRouteErrorHandler::buildWithCallback(function ($error, $code) {
+    if ($code == 404) {
+        echo "404!";
+    } else {
+        echo json_encode(['message' => $error, 'code' => $code]);
+    }
+}));
+
 $router->get("getDocument/{doc_id}/page/{page_id}", function ($docId, $pageId) {
     echo "GET DOC {$docId} PAGE {$pageId}" . PHP_EOL;
 });
@@ -36,4 +45,7 @@ $router->loadAllControllersInDirectoryAsCI(
         //'no_such_filter',//this might cause error
     ]
 );
+
+$router->frontendFolder("qd/ym", __DIR__ . '/frontend', []);
+
 $web_service->handleRequest();