+++ web/ArkRequestFilter.php
@@ -19,6 +19,7 @@ abstract class ArkRequestFilter
     {
         if (!empty($class_name)) {
             try {
+                /** @noinspection PhpIncompatibleReturnTypeInspection */
                 return class_exists($class_name) ? new $class_name() : self::generateFilter(true, 'Required a filter not existed.', 500);
             } catch (\Exception $exception) {
                 return self::generateFilter(true, 'Exception in requiring a filter: ' . $exception->getMessage(), 500);