+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Change Log of Ark
 
-### Coming Version 1.5.0
+### Version 1.5.0
+
+> Take attention when upgrading from 1.4.x or earlier versions.
 
 Parse the incoming user agent by `Jenssegers\Agent\Agent`.
 Auto detect MIME for Content Type Header by `Mimey\MimeTypes`.
+Make class parameter of `ArkRequestFilter` to be `string` or `string[]`.
+Rename router quick method `option` to be `options`.
 
 ### Version 1.4.2